Users
User authorization, Social Auth with Facebook (Socialite package installed), role permissions for admin, agent and customers. Notifications via email for validating accounts in registering.

+++ app/Http/Controllers/UsersController.php
@@ -5,6 +5,7 @@ namespace QuickUser\Http\Controllers;
 use Illuminate\Http\Request;
 use QuickUser\Http\Requests\CreateUserRequest;
 use QuickUser\Http\Requests\UpdateUserRequest;
+use QuickUser\Notifications\AdminUserCreated;
 use QuickUser\User;
 use QuickUser\Role;
 
@@ -12,40 +13,50 @@ class UsersController extends Controller
 {
   public function index()
   {
+    $this->authorize('access_users_list', User::class);
     $users = User::orderBy('created_at', 'desc')->paginate(10);
     return view('users.index', [ 'users' => $users ]);
   }
 
   public function profile(User $user)
   {
+    $this->authorize('update', [User::class, $user]);
     return view('users.edit', [ 'user' => $user, 'roles' => $this->get_all_roles() ]);
   }
 
   public function create()
   {
+    $this->authorize('create', User::class);
     return view('users.new', [ 'roles' => $this->get_all_roles() ]);
   }
 
   public function save(CreateUserRequest $request)
   {
+    $this->authorize('create', User::class);
+
+    $pass = $this->get_password($request);
     $user = User::create([
       'name' => $request->input('name'),
       'phone_number' => $request->input('phone_number'),
       'email' => $request->input('email'),
-      'password' => $this->get_password($request),
+      'password' => bcrypt($pass),
       'role_id' => $request->input('role_id', 3)
     ]);
 
-    return redirect()->route('users_path')->withSuccess('Usuario creado satisfactoriamente!!');
+    $user->notify(new AdminUserCreated($user, $pass));
+
+    return redirect()->route('active_user', $user->id);
   }
 
   public function edit(User $user)
   {
+    $this->authorize('update', [User::class, $user]);
     return view('users.edit', [ 'user' => $user, 'roles' => $this->get_all_roles() ]);
   }
 
   public function update(User $user, UpdateUserRequest $request)
   {
+    $this->authorize('update', [User::class, $user]);
     $user->update([
       'name' => $request->input('name'),
       'phone_number' => $request->input('phone_number'),
@@ -59,28 +70,39 @@ class UsersController extends Controller
 
   public function delete(User $user)
   {
+    $this->authorize('delete', User::class);
     $user->delete();
     return redirect()->route('users_path')->withSuccess('Usuario eliminado correctamente!!');
   }
 
   public function active_user(User $user)
   {
-    $user->actived = 1;
+    $this->authorize('active_user', User::class);
+    $user->active = 1;
     $user->save();
+
     return redirect()->route('edit_user_path', $user->id);
   }
 
   public function disable_user(User $user)
   {
-    $user->actived = 0;
+    $this->authorize('disable_user', User::class);
+    $user->active = 0;
     $user->save();
+
+    if($user->verify_if_current_user_account_is_disabled())
+    {
+      auth()->logout();
+      return redirect('/');
+    }
+
     return redirect()->route('edit_user_path', $user->id);
   }
 
   private function get_password(CreateUserRequest $request){
-    $password = bcrypt($request->input('password'));
+    $password = $request->input('password');
     if($request->has('check_random_pass')){
-      $password = bcrypt(str_random(10));
+      $password = str_random(10);
     }
 
     return $password;
